Keep users insertion order
Using object spread operator to override a user property in the state, causes the order of the keys to change (oddly enough, this only happens in safari using the production build), and the modified user goes to the last of the list. When using Object.assign instead, the keys order does not change.